% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{generate_colors}
\alias{generate_colors}
\title{Generate a vector of colours from a named palette}
\usage{
generate_colors(n, palette = "viridis", ...)
}
\arguments{
\item{n}{Number of colours to generate (positive integer).}

\item{palette}{Name of the colour palette to use. Can be a viridisLite palette
(\code{"viridis"}, \code{"magma"}, \code{"plasma"}, \code{"inferno"}, \code{"cividis"}, \code{"mako"},
\code{"rocket"}, \code{"turbo"}), a grDevices palette (\code{"hcl"}, \code{"rainbow"},
\code{"heat"}, \code{"terrain"}, \code{"topo"}), an RColorBrewer palette (e.g., \code{"Set1"},
\code{"Spectral"}, \code{"Dark2"}), or a function that generates colours.}

\item{...}{Additional arguments passed to the palette function.}
}
\value{
A character vector of colour hex codes.
}
\description{
Generate a vector of colours from a named palette
}
\examples{
generate_colors(5, "viridis")
generate_colors(3, "Spectral")
generate_colors(4, "Dark2")

}
