% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_schedule-package.R
\docType{package}
\name{PrintSchedule-package}
\alias{PrintSchedule}
\alias{PrintSchedule-package}
\title{PrintSchedule: Print Work Schedules from Shared Calendar Files}
\description{
A Shiny application for generating printable calendars from shared .ics calendar files (iCal format). Designed for healthcare workers to print work schedules from MyPlan shared calendar links. Features include color-coded event types, week and month views, and PDF export.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agdamsbo/print_schedule}
  \item Report bugs at \url{https://github.com/agdamsbo/print_schedule/issues}
}

}
\author{
\strong{Maintainer}: Andreas Gammelgaard Damsbo \email{andreas@gdamsbo.dk} (\href{https://orcid.org/0000-0002-7559-1154}{ORCID})

}
\keyword{internal}
