% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Schedule Shiny Application}
\usage{
run_app()
}
\value{
This function normally does not return; it invokes \code{\link[shiny:runApp]{shiny::runApp()}}
to start the app. It will return the value returned by the app's server
function if the app is programmatically exited with a value.
}
\description{
Launches the Shiny app for printing work schedules from shared .ics calendar files.
}
\examples{
\dontrun{
PrintSchedule::run_app()
}

}
